compliance-rag - 副本/requirements.txt
PyYAML
langchain
langchain-community
langchain-core
qwen
sentence-transformers
faiss-cpu
pydantic>=2.0